Use UMD bundling
Wraps the built library with UMD boilerplate. This should allow
consumers to import the AblePlayer using standard workflows while still
adding to the global object when loaded directly.

In order for people who use bundlers to have assets managed, the
specific paths and loading mechanisms needs to be controlled by the
calling code.

This is controlled by setting AbllPlayer.defaultOptions